Fixes and Enhancements in the Application
Addressed issues related to the display of 'Save' and 'Cancel' buttons on the food creation/editing page in the English locale.
Improved context handling to ensure the application works correctly with any context.
Added styles to the food creation/editing page.
The title on the food creation/editing page is now displayed unambiguously based on the situation ('Create meal' or 'Edit meal').
Resolved the problem with style rendering and food operations after filtering.
Moved the JspMealController to the 'meal' package.
Utilized aliases such as @GetMapping and @PostMapping to improve code readability.
Data is now passed to the Model abstraction in the controller.
Removed hardcoded values to make the application compatible with any context.
Added a CharacterEncodingFilter to ensure proper UTF-8 encoding.
Made improvements to the organization of controller code